🔥(base) removed storage and cache from site-factory
Removed storage and cache from site-factory and applied to Richie
in order to avoid repeating unnecessary and critical implementations
for each instance. If any customization is necessary, it is up to the
site-factory implementation. Otherwise, as expected, it keeps up to Richie.
