+++ cookiecutter/{{cookiecutter.organization}}-richie-site-factory/template/{{cookiecutter.site}}/src/backend/{{cookiecutter.site}}/settings.py
@@ -9,7 +9,7 @@ from django.utils.translation import gettext_lazy as _
 
 # pylint: disable=ungrouped-imports
 import sentry_sdk
-from base.utils import merge_dict
+from richie.apps.core.utils import merge_dict
 from configurations import Configuration, values
 from sentry_sdk.integrations.django import DjangoIntegration
 
@@ -201,7 +201,7 @@ class Base(StyleguideMixin, DRFMixin, RichieCoursesConfigurationMixin, Configura
             "BACKEND": "django.core.files.storage.FileSystemStorage",
         },
         "staticfiles": {
-            "BACKEND": "base.storage.CDNManifestStaticFilesStorage",
+            "BACKEND": "richie.apps.core.storage.CDNManifestStaticFilesStorage",
         },
     }
 
@@ -601,7 +601,7 @@ class Base(StyleguideMixin, DRFMixin, RichieCoursesConfigurationMixin, Configura
         {
             "default": {
                 "BACKEND": values.Value(
-                    "base.cache.RedisCacheWithFallback",
+                    "richie.apps.core.cache.RedisCacheWithFallback",
                     environ_name="CACHE_DEFAULT_BACKEND",
                     environ_prefix=None,
                 ),
@@ -776,10 +776,10 @@ class Production(Base):
 
     STORAGES = {
         "default": {
-            "BACKEND": "base.storage.MediaStorage",
+            "BACKEND": "richie.apps.core.storage.MediaStorage",
         },
         "staticfiles": {
-            "BACKEND": "base.storage.CDNManifestStaticFilesStorage",
+            "BACKEND": "richie.apps.core.storage.CDNManifestStaticFilesStorage",
         },
     }
     AWS_DEFAULT_ACL = None
