+++ src/richie/apps/core/storage.py
@@ -19,7 +19,7 @@ class CDNManifestStaticFilesStorage(ManifestStaticFilesStorage):
     and ignores files that are already versioned by webpack.
     """
 
-    def post_process(self, paths, dry_run=False, **options):
+    def post_process(self, paths, dry_run=False, **options):  # pylint: disable=W0221
         """
         Remove paths from file to post process.
         Some js static files generated by webpack already have a unique name per build
@@ -55,7 +55,7 @@ class CDNManifestStaticFilesStorage(ManifestStaticFilesStorage):
         return url
 
 
-class MediaStorage(S3Boto3Storage):
+class MediaStorage(S3Boto3Storage):  # pylint: disable=W0223
     """A S3Boto3Storage backend to serve media files via CloudFront."""
 
     bucket_name = getattr(settings, "AWS_MEDIA_BUCKET_NAME", None)
