+++ src/richie/apps/core/tests/test_settings.py
